requirements.txt
django
djangorestframework
django-cors-headers
python-dotenv
requests
msal
google-api-python-client
google-auth
pytz